requirements.txt
aiosqlite==0.20.0
alembic==1.14.0
annotated-types==0.7.0
anyio==4.5.2
async-timeout==5.0.1
asyncpg==0.30.0
bcrypt==3.2.2
cffi==1.17.1
click==8.1.8
dnspython==2.6.1
ecdsa==0.19.0
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.6
greenlet==3.1.1
h11==0.14.0
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
Mako==1.3.8
MarkupSafe==2.1.5
orjson==3.10.12
passlib==1.7.4
pyasn1==0.6.1
pycparser==2.22
pydantic==2.10.4
pydantic_core==2.27.2
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.20
rsa==4.9
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
typing_extensions==4.12.2
uvicorn==0.33.0
zipp==3.20.2
